+++ cognite/src/api/mod.rs
@@ -4,9 +4,9 @@ pub mod auth;
 pub mod authenticator;
 pub mod core;
 pub mod data_ingestion;
+pub mod data_modeling;
 pub mod data_organization;
 pub mod iam;
-pub mod models;
 pub mod resource;
 pub mod utils;
