--- cognite/src/dto/models/mod.rs
@@ -1,1 +1,0 @@
-pub mod instances;
