+++ cognite/src/lib.rs
@@ -62,11 +62,6 @@ pub mod sequences {
     pub use super::dto::core::sequences::*;
 }
 
-pub mod models {
-    pub use super::api::models::*;
-    pub use super::dto::models::*;
-}
-
 pub mod iam {
     pub use super::api::iam::{
         api_keys::*, groups::*, security_categories::*, service_accounts::*, sessions::*,
